Add user state-related features, Logout Service, Image Service, and get 'Main@Notice' / 'Work@Real-time Log', 'Work@Approval Log' init data

+++ Coconet-main/src/main/java/com/coconet/server/entity/ChartData.java
@@ -22,6 +22,7 @@ public class ChartData {
     private int num;
 
     private String title;
-    private int value;
+    private double value;
     private String color;
+
 }
